Add saddle find, format some code, add random fill

+++ src/operators.c
@@ -7,15 +7,15 @@
 extern int errno;
 
 Matrix add(const Matrix *m1, const Matrix *m2) {
-    if((m1->row != m2->row) || (m1->col != m2->col)) {
+    if ((m1->row != m2->row) || (m1->col != m2->col)) {
         errno = EINVAL;
         perror("Matrices must be the same size");
         return *m1;
     }
     Matrix ret;
     ma_alloc(&ret,m1->col,m1->row);
-    for(int i = 0; i < m1->row; i++) {
-        for(int j = 0; j < m1->col; j++) {
+    for (size_t i = 0; i < m1->row; i++) {
+        for (size_t j = 0; j < m1->col; j++) {
             ret.array[i][j] = 
                 m1->array[i][j] + m2->array[i][j];
         }
@@ -25,15 +25,15 @@ Matrix add(const Matrix *m1, const Matrix *m2) {
 } 
 
 Matrix subtract(const Matrix *m1, const Matrix *m2) {
-    if((m1->row != m2->row) || (m1->col != m2->row)) {
+    if ((m1->row != m2->row) || (m1->col != m2->row)) {
         errno = EINVAL;
         perror("Matrices must be the same size");
         return *m1;
     }
     Matrix ret;
     ma_alloc(&ret,m1->row, m1->col);
-    for(int i = 0; i < m2->row; i++)
-        for(int j = 0; j < m2->col; j++)
+    for (size_t i = 0; i < m2->row; i++)
+        for (size_t j = 0; j < m2->col; j++)
             ret.array[i][j] = 
                 m1->array[i][j] - m2->array[i][j];
     errno = 0;
@@ -41,7 +41,7 @@ Matrix subtract(const Matrix *m1, const Matrix *m2) {
 }
 
 Matrix multiply(const Matrix *m1, const Matrix *m2) {
-    if(m1->col != m2->row) {
+    if (m1->col != m2->row) {
         errno = EINVAL;
         perror("Column of matrix 1 must be equal to column of matrix 2");
         return *m1;
@@ -49,9 +49,9 @@ Matrix multiply(const Matrix *m1, const Matrix *m2) {
     Matrix ret;
     ma_alloc(&ret,m1->row,m2->col);
 
-    for(int i = 0; i < m1->row; ++i) {
-        for(int j = 0; j < m2->col; ++j) {
-            for(int k = 0; k < m1->row; ++k) {
+    for (size_t i = 0; i < m1->row; ++i) {
+        for (size_t j = 0; j < m2->col; ++j) {
+            for (size_t k = 0; k < m1->row; ++k) {
                 ret.array[j][k] += 
                     m1->array[j][i] * m2->array[i][k];
             }
@@ -70,9 +70,9 @@ Matrix divide(const Matrix *m1, const Matrix *m2) {
     Matrix ident = identityMatrix(m2->row);
     Matrix ret = cat(m2, &ident,0);
     float factor = 1/det(m2); 
-    for(int i = 0; i < m2->row; i++) { 
-        for(int j = 0; j < m2->col; j++) { 
-            if( (j+1+i) % 2 == 0) 
+    for (size_t i = 0; i < m2->row; i++) { 
+        for (size_t j = 0; j < m2->col; j++) { 
+            if ( (j+1+i) % 2 == 0) 
                 m2->array[i][j] *= -1; 
         } 
     }
